% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_mean.R
\name{basic_mean}
\alias{basic_mean}
\title{A basic mean calculator}
\usage{
basic_mean(vec)
}
\arguments{
\item{vec}{A numeric vector}
}
\value{
A double
}
\description{
A basic mean calculator
}
\examples{
basic_mean(1:3)
}
