% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexgamma.R
\name{cgamma}
\alias{cgamma}
\title{R interface to omplex gamma function of a vector or matrix argument.}
\usage{
cgamma(z)
}
\arguments{
\item{z}{A vector or matrix (numeric or complex)}
}
\value{
Returns a vector or matrix of complex values of the gamma function
 corresponding to the input numeric or complex values.
}
\description{
R interface to omplex gamma function of a vector or matrix argument.
}
\details{
Computes the complex gamma function using the Lanczos approximation
 (see Wikipedia).
}
\examples{
m <- matrix(1:12,3,4)
cgamma(m)

}
