% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cgammacpp}
\alias{cgammacpp}
\title{R interface to complex gamma function.}
\usage{
cgammacpp(inRvec)
}
\arguments{
\item{inRvec}{complex vector or 2d matrix of complex numbers}
}
\value{
Returns a vector or matrix of complex values.
}
\description{
R interface to complex gamma function.
}
\details{
The `Rcomplex` data structure definition has changed
recently in `R_ext/Complex.h`.
}
