% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_plot.R
\name{corr_heatmap}
\alias{corr_heatmap}
\title{Plot the correlation heatmap}
\usage{
corr_heatmap(df)
}
\arguments{
\item{df}{A data frame containing medical insurance data.}
}
\value{
A correlation heatmap plot
}
\description{
Plot the correlation heatmap
}
\details{
This function allows the user to plot a correlation heatmap to see the correlations
among the variables of the insurance dataframe
}
