% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpptoolchain}
\alias{cpptoolchain}
\title{Return C++ toolchain used to build package \code{Mypack}}
\usage{
cpptoolchain()
}
\description{
Return C++ toolchain used to build package \code{Mypack}
}
\details{
The Windows toolchain (MSVC) shows version info MMNNBBBBB, where
MM=Major, NN=Minor, and BBBBB=Build.
}
