% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groundwater_depth.R
\name{depth}
\alias{depth}
\title{Depth to groundwater}
\usage{
depth(groundwater_depth)
}
\arguments{
\item{depth_groundwater, }{depth to groundwater in meters m, numeric variable}

\item{n, }{year since year 0, numeric variable}
}
\value{
depth_groundwater, numeric variable
}
\description{
Function to calculate depth to groundwater at a specific year, knowing the initial depth to groundwater of 10m at year 0 and an increase in depth by 10% each year.
}
\author{
Annika Leiby
}
