% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_corr_bounds.R
\name{estimate_corr_bounds}
\alias{estimate_corr_bounds}
\title{Estimate bounds of indirect correlations}
\usage{
estimate_corr_bounds(corrMat)
}
\arguments{
\item{corrMat}{Matrix Object which represents a correlation matrix. Indirect correlations must be set to
\code{NA} within the matrix. This matrix must contain at least two \code{NA} values (it is symmetric).}
}
\value{
A matrix object with N(= number of indirect correlations) rows and 4 columns:
\itemize{
   \item{\code{var1}:} {numerical index of the first variable of the indirect correlation couple}
   \item{\code{var2}:} {numerical index of the second variable of the indirect correlation couple}
   \item{\code{lower}:} {lower bound for the range of indirect correlation between \code{var1} and \code{var2}}
   \item{\code{upper}:} {upper bound for the range of indirect correlation between \code{var1} and \code{var2}}
   }
}
\description{
For each indirect correlation to be estimated, this function returns the range of its possible values.
Indirect correlations must be identified in input correlation matrix with \code{NA} values. From input
correlation structure is derived the associated graph. Then, for each indirect correlation between the generic
couple of variables (\eqn{X1},\eqn{X2}),  all the possible paths in the graph that links \eqn{X1} and \eqn{X2}
visiting one node at most once are considered.

For each path is computed its cost by multiplying the correlations along it, range of indirect correlation values is
taken considering the minumum and the maximum costs. If does not exist any path betwen two nodes, maximum and minimum values will be set to \code{NA}. If
there is a unique path, the bound is computed considering \eqn{cost +- 0.5*cost}.
}
\examples{
# create a correlation matrix, define some correlations
c_start <- diag(rep(1,6))
c_start[1,2] <- -0.6
c_start[1,3] <- -0.75
c_start[2,3] <-0.95
c_start[2,4] <- 0.75
c_start[2,6] <- -0.6
c_start <- c_start+t(c_start)-diag(rep(1,ncol(c_start)))
#set to NA indirect correlations
c_start[c_start==0]<-NA
#plot correlation graph
plot_graph_corr(c_start,"Graph of Correlation Matrix")
#get bounds of correlations
estimate_corr_bounds(c_start)

#output: variable 5 is not directly correlated with any of the others so it
#is impossible to establish a path to it. So, indirect correlations bounds with variable 5 are set to NA

#       var1 var2     lower   upper
#[1,]    1    4 -0.534375 -0.4500
#[2,]    1    5        NA      NA
#[3,]    1    6  0.360000  0.4275
#[4,]    2    5        NA      NA
#[5,]    3    4  0.337500  0.7125
#[6,]    3    5        NA      NA
#[7,]    3    6 -0.570000 -0.2700
#[8,]    4    5        NA      NA
#[9,]    4    6 -0.540000 -0.3600
#[10,]    5    6        NA      NA
}
\seealso{
\code{\link[igraph]{graph_from_adjacency_matrix}}

\code{\link[igraph]{all_simple_paths}}

\code{\link{get_pathProduct}}

\code{\link{get_bounds}}
}
\author{
Alessandro De Carlo \email{alessandro.decarlo01@universitadipavia.it}
}
