% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{factoring}
\alias{factoring}
\title{Easily convert the character columns to factors in medical insurance data}
\usage{
factoring(df)
}
\arguments{
\item{df}{A data frame containing medical insurance data.}
}
\value{
A data frame
}
\description{
Easily convert the character columns to factors in medical insurance data
}
\details{
This function allows the user to convert all the columns with type character to factor
to be able to use it for prediction, because factors have different levels like numbers, so it
is somehow like encoding the categorical variables
}
\examples{
factoring(insurance)
}
