% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This is a function that plots the LONGITUD and LATITUDE of specified state on a map, whcih has to appear in the data frame specified by a ceetain year.(using the \code{state.num,year} argument).}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number giving the id of a state. This function
reads in \code{state.num} and \code{year}. In the file read accordin
to the \code{year}, there may not be matched \code{state.num}, and
this case the function stops and prints "invalid STATE number: ",
\code{state.num}). The filtered data according to \code{state.num}
and \code{year} may not exist, and in this case it prints a warning
message "no accidents to plot" and return a NULL object.}

\item{year}{A number giving the year}
}
\value{
This function plots the LONGITUD and LATITUDE of the selected states on a map.
}
\description{
The function "filter" is imported from external package "dplyr".
The function "map" is imported from external package "maps".
}
\examples{
fars_map_state(25, 2015)

}
