% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{This is a function that reads in a data file (using the \code{filename} #'argument).}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of a csv file. When #'the function can not find the file, it stops and prints an error message #'"file \code{filename} does not exist".}
}
\value{
This function returns a data frame
}
\description{
The function "read_csv" is imported from external package "readr"
The function "tbl_df" is imported from external package "dplyr"
}
\examples{
fars_read("data.csv")

}
