% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{This is a function that selects files containing a specified list of #'years, and in each data frame representing data of each selected file #'creats a new year columne, and finally select this new conlumn and the #'MONTH column.(using the \code{years} argument).
The function "mutate" and "select" is imported from external package #'"dplyr"}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list containing years to select. This function calls #'"fars_read" function to read files corresponding to the given years. For #'a given year, the file may not exist, and then will prints an error #'message "invalid year: ", \code{year} and return a NULL object for this #'year.}
}
\value{
This function returns a data frame containing the year column #'and the month column. As a side effect, this function also prints out
   these two columns.
}
\description{
This is a function that selects files containing a specified list of #'years, and in each data frame representing data of each selected file #'creats a new year columne, and finally select this new conlumn and the #'MONTH column.(using the \code{years} argument).
The function "mutate" and "select" is imported from external package #'"dplyr"
}
\examples{
fars_read_years(list(2014, 2013, 2011))

}
