% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{This is a function that selects files containing a specified list of #'years, and then select specific columns containing year and month, and #'row-combines these columns from each file before summarizing the result #'by counting the same year and month pair and spread the year and count #'into multiple columns(using the \code{years} argument).}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list containing years to select}
}
\value{
This function returns a summary with year and counts of each #'month appearances
}
\description{
The function "bind_rows", "group_by" and "summarizeis" are imported from #'external package "dplyr".
The function "spread" is imported from external package "tidyr".
the piping operator "%>%" requires external package "magrittr".
}
\examples{

WD <- setwd(system.file("extdata", package = "mypack"))
fars_summarize_years(2013)
fars_summarize_years(c(2013, 2014))
setwd(WD)
}
