% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_data.R
\docType{data}
\name{fish_data}
\alias{fish_data}
\title{Fish Species Vector}
\format{A data frame with 300 rows and 1 colums
\describe{
\item{x}{character, salmon, steelhead, tuna, shark, cod}
}}
\usage{
fish_data
}
\description{
Vector length of 300 with random pattern of species
}
\keyword{datasets}
