% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_fun.R
\name{fish_fun}
\alias{fish_fun}
\title{Analyze fish}
\usage{
fish_fun(fish, plot = FALSE)
}
\arguments{
\item{fish}{the vector of fish}

\item{plot}{whether to plot histogram or not; default is FALSE}
}
\value{
a list containing the most common fish, the rarest fish, the total number of fish, and a histogram if asked
}
\description{
function returns the most common fish, the rarest fish, the total number of fish, and a histogram if asked
}
\author{
Shuhan Song and Annika Leiby
}
