% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bounds.R
\name{get_bounds}
\alias{get_bounds}
\title{Compute the maximum and the minimun cost considering all the possible paths between two nodes in correlation
graph}
\usage{
get_bounds(corrmat, list_path)
}
\arguments{
\item{corrmat}{Matrix Object. It represents the correlation structure among variables}

\item{list_path}{List Object. It contains all the possible paths between two nodes of the graph.}
}
\value{
Array object with minimun and maximun cost considering the costs of all the possible paths between a couple
of nodes
}
\description{
Given a correlation graph, a starting node and a target node, this function takes in input the list
of all possible paths which link the starting node and the target visiting each node at most once. For each path, it is computed the relative
cost by multiplying the correlations met along the path. Then, their minimum and maximum value are returned in a 2-d
array. If does not exist any path betwen two nodes, maximum and minimum values will be set to \code{NA}. If
there is a unique path, the bound is computed considering \eqn{cost +- 0.5*cost}.
}
\note{
This function is used in the workflow implemented for estimating indirect correlations for computing the bounds
values to estimate
}
\seealso{
\code{\link{get_pathProduct}}
}
\author{
Alessandro De Carlo \email{alessandro.decarlo01@universitadipavia.it}
}
