% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corrMatrixOptim.R
\name{get_corrMatrixOptim}
\alias{get_corrMatrixOptim}
\title{Get final correlation matrix with estimated indirect correlations}
\usage{
get_corrMatrixOptim(x_opt, cbase, var_optim)
}
\arguments{
\item{x_opt}{vector with estimated indirect correlations matrix.}

\item{cbase}{Matrix object. It represents the input correlation matrix. Indirect correlations to be estimated must be
explicitated with NA in this matrix.}

\item{var_optim}{Matrix object. It  has a row number equal to the number of indirect correlations to be estimated and
and 4 columns:
\itemize{
   \item{\code{var1}:} {numerical index of the first variable of the indirect correlation couple}
   \item{\code{var2}:} {numerical index of the second variable of the indirect correlation couple}
   \item{\code{lower}:} {lower bound for the range of indirect correlation between \code{var1} and \code{var2}}
   \item{\code{upper}:} {upper bound for the range of indirect correlation between \code{var1} and \code{var2}}
   }}
}
\value{
Matrix object which contains the original correlation matrix with indirect correlations estimated.
}
\description{
This function returns the final correlations matrix with the estimated values of indirect correlations.
}
\seealso{
\code{\link{optim.fun}}
}
\author{
Alessandro De Carlo \email{alessandro.decarlo01@universitadipavia.it}
}
