% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathProduct.R
\name{get_pathProduct}
\alias{get_pathProduct}
\title{Compute the product of all correlations along a path}
\usage{
get_pathProduct(corrMat, nodes)
}
\arguments{
\item{corrMat}{Matrix Object. It represent the correlation matrix}

\item{nodes}{Vector. It contains all the nodes of the path that links \code{nodes[1]} and \code{nodes[length(nodes)-1]}}
}
\value{
Scalar double which represents the product of all correlations along a path between two nodes in the
correlations graph.
}
\description{
This function returns the product of all correlations along a path between two nodes in the correlations
graph.
}
\author{
Alessandro De Carlo \email{alessandro.decarlo01@universitadipavia.it}
}
