% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groundwater_category.R
\name{gw_category}
\alias{gw_category}
\title{Groundwater Category}
\usage{
gw_category(level)
}
\arguments{
\item{level}{water level in meters, numeric variable}
}
\value{
category based on the water level
}
\description{
This function categorizes a well as "sustainable", "overdrafted", or "critically overdrafted" based on water level
}
\author{
Annika Leiby
}
