% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insurance}
\alias{insurance}
\title{(Dataset)Medical cost insurance data}
\format{
## `who`

A data frame with 1338 rows and 7 columns:
\describe{
  \item{age}{age of primary beneficiary}
  \item{sex}{insurance contractor gender, female, male}
  \item{bmi}{Body mass index, providing an understanding of body, weights that are relatively high or low relative to height}
  \item{children}{Number of children covered by health insurance / Number of dependents}
  \item{smoker}{smoking}
  \item{region}{the beneficiary's residential area in the US, northeast, southeast, southwest, northwest.}
  \item{charges}{Individual medical costs billed by health insurance}
  ...
}
}
\usage{
insurance
}
\description{
A tidy dataset of medical cost by the insurance company
}
\keyword{datasets}
