% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llqr.R
% 描述：该函数使用 rq 函数拟合局部线性分位数回归。
\name{llqr}
\alias{llqr}
\title{Local linear Quantile Regression}
\usage{
llqr(x, y, h, tau, m = 200, kernel = FALSE)
}
\arguments{
\item{x}{The design matrix}

\item{y}{The response variable}

\item{h}{The bandwidth parameter of a kernel function}

\item{tau}{The tau quantile (0 to 1)}

\item{m}{The number of points at which the function is to be estimated}

\item{kernel}{The kernel function with several types. \code{norm} is to chose the Normal kernal}
}
\value{
parameter estimation at sample points.
}
\description{
fit a local linear quantile regression with rq function,
}
