% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{This is a function that prints a file name containing a specific year
number.(using the \code{year} argument).}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number giving the year}
}
\value{
This function prints and returns a string file name
}
\description{
This is a function that prints a file name containing a specific year
number.(using the \code{year} argument).
}
\examples{

WD <- setwd(system.file("extdata", package = "mypack"))
make_filename(2013)
setwd(WD)
}
