% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max.R
\name{min_max}
\alias{min_max}
\title{Calcula minimo y maximo de un vector aleatorio}
\usage{
min_max(n, media, desvi)
}
\arguments{
\item{n}{Tamaño de la muestra.}

\item{media}{media de la población}

\item{desvi}{desviación de la población}
}
\value{
un vector con mínimo y máximo
}
\description{
Calcula minimo y maximo de un vector aleatorio
}
\examples{
min_max(n=10,media=50, desvi=2)
min_max(n=100,desvi=3, media=180)

}
