\name{mypack-package}
\alias{mypack-package}
\alias{mypack}
\docType{package}
\title{
\packageTitle{mypack}
}
\description{
\packageDescription{mypack}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{mypack}

Maintainer: \packageMaintainer{mypack}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = mypack)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{mypack}}
%\section{Documentation Index}{\packageIndices{mypack}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
