% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyWeibull.R
\name{MyWcdf}
\alias{MyWcdf}
\title{CDF of Weibull distribution}
\usage{
MyWcdf(x, a, b)
}
\arguments{
\item{x}{value.}

\item{a}{Shape parameter.}

\item{b}{Scale parameter.}
}
\value{
Cumulative density.
}
\description{
This function calculates the cumulative density for x given ‘a’
as the shape parameter and ‘b’as the scale parameter.
}
