% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyWeibull.R
\name{MyWpdf}
\alias{MyWpdf}
\title{PDF of Weibull distribution}
\usage{
MyWpdf(x, a, b)
}
\arguments{
\item{x}{value.}

\item{a}{Shape parameter.}

\item{b}{Scale parameter.}
}
\value{
Probability density.
}
\description{
This function calculates the probability density for x given ‘a’
as the shape parameter and ‘b’as the scale parameter.
}
