% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimfun.R
\name{optim.fun}
\alias{optim.fun}
\title{Optimization function}
\usage{
optim.fun(x, cbase, var_optim)
}
\arguments{
\item{x}{vector of unknown variables (indirect correlations)}

\item{cbase}{Matrix object. It represents the input correlation matrix. Indirect correlations to be estimated must be
explicitated with NA in this matrix.}

\item{var_optim}{Matrix object. It  has a row number equal to the number of indirect correlations to be estimated and
and 4 columns:
\itemize{
   \item{\code{var1}:} {numerical index of the first variable of the indirect correlation couple}
   \item{\code{var2}:} {numerical index of the second variable of the indirect correlation couple}
   \item{\code{lower}:} {lower bound for the range of indirect correlation between \code{var1} and \code{var2}}
   \item{\code{upper}:} {upper bound for the range of indirect correlation between \code{var1} and \code{var2}}
   }}
}
\value{
The negative of the minimum eigenvalue of correlation matrix. This represents the quantity to minimize in
the optimization problem for estimating indirect correlations.
}
\description{
This function provide a general implementation of the functional that has to be minimized in order to
estimate indierect correlations. Given the starting correlation matrix with the indirect correlations to be estimated
identified by NA, it automatically generate a matrix which depends by a vector, x, of unknown variables which
represents indirect correlations. During the optimizazion task, in order to obtain a PD correlation matrix, the task
is to minmize the negative of the minimum eigenvalue in this matrix.
}
\author{
Alessandro De Carlo \email{alessandro.decarlo01@universitadipavia.it}
}
