% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_plot.R
\name{pred_plot}
\alias{pred_plot}
\title{Plot the predicted values vs. the actual values}
\usage{
pred_plot(df)
}
\arguments{
\item{df}{A data frame containing medical insurance data.}
}
\value{
A plot
}
\description{
Plot the predicted values vs. the actual values
}
\details{
This function allows the users to plot the actual values vs. the predicted values
}
