% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{predicted}
\alias{predicted}
\title{Predict the charges using Random Forest model}
\usage{
predicted(df, model = rf)
}
\arguments{
\item{df}{A data frame containing medical insurance data.}

\item{model}{The model which is used for prediction}
}
\value{
A random forets model summary
}
\description{
Predict the charges using Random Forest model
}
\details{
This function allows the users to find the predicted values for charges using the random forest model
}
