% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_Forest.R
\name{rf}
\alias{rf}
\title{Train a random forest model on charges}
\usage{
rf(df)
}
\arguments{
\item{df}{A data frame containing medical insurance data.}
}
\value{
A random forets model summary
}
\description{
Train a random forest model on charges
}
\details{
This function allows the users to train a random forest model on medical insurance
 dataframe to predict how much charges the insurance company should pay
}
