% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_drv.R
\name{rw.check.drv}
\alias{rw.check.drv}
\title{Check equality between two variables}
\usage{
rw.check.drv(x0, x1, db, var.info)
}
\arguments{
\item{x0}{A variable to be checked}

\item{x1}{The variable used to check x0. If x1 is not provided, <x0>_RW will be used. 
For example, if x0 = VAR, then VAR_RW will be the default for comparison.}

\item{db}{A dataset}

\item{var.info}{The variables to be displayed along with x0 and x1 for information}
}
\value{
A dataset showing records with non-identical values
}
\description{
Check equality between two variables
}
