% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_tlf.R
\name{rw.compare.tlf}
\alias{rw.compare.tlf}
\title{TLF Output Comparison}
\usage{
rw.compare.tlf(env = "VST", path.new, path.old, path.out)
}
\arguments{
\item{env}{The environment where the outputs are generated. Default: VST}

\item{path.new}{The full directory path where the new outputs are located.}

\item{path.old}{The full directory path where the old outputs are located.}

\item{path.out}{The full directory path for saving the list of changed outputs.}
}
\value{
A PDF showing outputs that differ between versions
}
\description{
TLF Output Comparison
}
