% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{rw.date}
\alias{rw.date}
\title{Impute a partial date}
\usage{
rw.date(x, format = "ymd", m, d, date.ref)
}
\arguments{
\item{x}{The character string of date to be imputed}

\item{format}{A character string, specifying the date formats, including 'ymd' (default), 'dmy', 'mdy'}

\item{m}{A numeric value used to impute the month}

\item{d}{A numeric value used to impute the day}

\item{date.ref}{The reference date, in the same date format, for comparison with the imputed date. 
The reference date is displayed if the available components of the partial date match 
the corresponding components of the reference date}
}
\value{
An imputed date
}
\description{
Impute a partial date
}
