% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_desc.R
\name{rw.desc}
\alias{rw.desc}
\title{Descriptive Summary for Numeric Variables}
\usage{
rw.desc(db, trt, aval, l1, l2, stats = "common", dp = 0, direction = "long")
}
\arguments{
\item{db}{A dataset, e.g., ADSL}

\item{trt}{A treatment variable, e.g., 'TRTA'}

\item{aval}{The numeric variable to be summarized, e.g., AVAL}

\item{l1}{Level 1, e.g., 'AVISIT'}

\item{l2}{Level 2, e.g., 'PARAMCD'}

\item{stats}{Character. Which stats to produce. Either “common” (default), or a selection of : 
'n', 'mean', 'sd', 'mean_sd', 'med', 'min', 'max', 'range', 'q1', 'q3', 'cv'.}

\item{dp}{Decimal place to be displayed}

\item{direction}{Display direction: long or wide}
}
\value{
A dataset with common stats, including n, mean, sd, min, max
}
\description{
Descriptive Summary for Numeric Variables
}
