% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_freq.R
\name{rw.freq}
\alias{rw.freq}
\title{Frequency Summary for Categorical Variables}
\usage{
rw.freq(
  db,
  trt,
  l1,
  l2,
  l3,
  N = "n",
  stats = "n_pct",
  ci.method = "clopper-pearson",
  conf.level = 0.95,
  diffci.method,
  diff.reff
)
}
\arguments{
\item{db}{A dataset, e.g., ADSL}

\item{trt}{A treatment variable, e.g., 'TRTA'}

\item{l1}{Level 1, e.g., 'AVALC'}

\item{l2}{Level 2, e.g., 'AVISIT'}

\item{l3}{Level 3, e.g., 'PARAMCD'}

\item{N}{The sample size serves as the denominator in the percentage calculation. 'n' for observed records (default)); 
a table or named vector to specify the n for each group, e.g., c('Group A' = 100, 'Group B' = 1000)}

\item{stats}{Character. Which stats to produce. Either “n_pct” (default), or a selection of : 
'n', 'n_n', 'n_pct', 'pct', 'ci', 'diff', 'diffci'}

\item{ci.method}{A character string, specifying the method for CI from BinomCI; e.g., "clopper-pearson", "wald", "wilson", "logit"}

\item{conf.level}{The confidence level, defaults to 0.95}

\item{diffci.method}{# A character string, specifying the method for CI of diff from BinomDiffCI;
e.g., 'wald' for traditional Wald; 'score' for Newcombe; 'scorecc' for Newcombe with continuity correction}

\item{diff.reff}{A character string, specify the trt group to be the reference for comparison}
}
\value{
A dataset with frequency counts, CI, diff, and/or CI of diff
}
\description{
Frequency Summary for Categorical Variables
}
