% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_freq1.R
\name{rw.freq1}
\alias{rw.freq1}
\title{Frequency Summary for Categorical Variables, taking unique records}
\usage{
rw.freq1(
  db,
  subj,
  trt,
  l1,
  l2,
  N,
  stats = "n",
  sort.l1 = "freq",
  sort.l2 = "freq",
  l1.desc = TRUE,
  l2.desc = TRUE,
  sort.ref = "OVERALL",
  diffci.method,
  diff.ref,
  conf.level = 0.95
)
}
\arguments{
\item{db}{A dataset, e.g., ADAE}

\item{subj}{A subject ID variable, e.g., USUBJID}

\item{trt}{A treatment variable, e.g., 'TRTA'}

\item{l1}{Level 1, e.g., 'PT'}

\item{l2}{Level 2, e.g., 'SOC'}

\item{N}{The sample size serves as the denominator in the percentage calculation; 
a table or named vector to specify the n for each group, e.g., c('Group A' = 100, 'Group B' = 1000)#'}

\item{stats}{Character. Which stats to produce. Either “n_pct” (default), or a selection of : 
'n', 'n_pct', 'diff', 'diffci'}

\item{sort.l1}{The sorting method to be applied to l1: 'freq' for frequency (default); 'alphabet' for alphabetical}

\item{sort.l2}{The sorting method to be applied to l2: 'freq' for frequency (default); 'alphabet' for alphabetical}

\item{l1.desc}{A logical specifying whether the sorting of l1 is performed by descending (default = TRUE) or not}

\item{l2.desc}{A logical specifying whether the sorting of l2 is performed by descending (default = TRUE) or not}

\item{sort.ref}{A character string, specifying the trt group to be the reference for sorting}

\item{diffci.method}{# A character string, specifying the method for CI of diff from BinomDiffCI;
e.g., 'wald' for traditional Wald; 'score' for Newcombe; 'scorecc' for Newcombe with continuity correction}

\item{diff.ref}{A character string, specify the trt group to be the reference for comparison}

\item{conf.level}{The confidence level, defaults to 0.95}
}
\value{
A dataset with unique frequency counts diff, and/or CI of diff
}
\description{
Frequency Summary for Categorical Variables, taking unique records
}
