% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sdtm.R
\name{rw.merge.sdtm}
\alias{rw.merge.sdtm}
\title{Merge SDTM Main and SUPP Domains}
\usage{
rw.merge.sdtm(sdtm, supp)
}
\arguments{
\item{sdtm}{A SDTM main domain}

\item{supp}{A SDTM SUPP domain}
}
\value{
A merged data.table with all variables from the main domain as well as the variables from all QNAM values
}
\description{
Merge SDTM Main and SUPP Domains
}
