% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\name{rw.paste}
\alias{rw.paste}
\title{Combine multiple columns into one by pasting strings together}
\usage{
rw.paste(.SD, collapse = " ")
}
\arguments{
\item{.SD}{A dataset/data.table}

\item{collapse}{A character string to separate the results}
}
\value{
A character vector of the concatenated values after removing NA and ''
}
\description{
Combine multiple columns into one by pasting strings together
}
\examples{
# Used in the data.table
# dbin[, .(NEW = rw.paste(.SD[, c('VAR1', 'VAR2', 'VAR3')], collapse = ''))]

}
