% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Plot the scatter plot}
\usage{
scatter_plot(df, col_of_interest_numeric, col_of_interest_factor)
}
\arguments{
\item{df}{A data frame containing medical insurance data.(called "insurance")}

\item{col_of_interest_numeric}{A numeric column for the x axis, the y axis is the target variable(It must be given with quotes)}

\item{col_of_interest_factor}{A data frame containing medical insurance data.(It must be given with quotes)}
}
\value{
A scatter plot
}
\description{
Plot the scatter plot
}
\details{
This function lets the user to have a scatter plot between 2 numeric variables(one is the target variable called charges and one taken from the user)
and fill with 1 categorical variable(taken from the user)
}
