% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexgamma.R
\name{showgamma}
\alias{showgamma}
\title{Shows 3D plot of Complex Gamma Function}
\usage{
showgamma()
}
\description{
Shows 3D plot of Complex Gamma Function
}
\details{
When used with CRcpp framework be sure to use x11() to
bring up a graphics window before invoking this function.
}
\examples{
showgamma()

}
