% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splitter.R
\name{split_data}
\alias{split_data}
\title{Split the dataframe to train and test}
\usage{
split_data(df)
}
\arguments{
\item{df}{A data frame containing medical insurance data.}
}
\value{
A dataframe of 2 columns train and test
}
\description{
Split the dataframe to train and test
}
\details{
This function allows the users to have split of the dataframe
to train set and test set for machine learning tasks
}
