% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Funcion para sumar}
\usage{
suma(a, b)
}
\arguments{
\item{a}{primer número}

\item{b}{segundo número}
}
\value{
muestra el resultado de la suma
}
\description{
Funcion para sumar
}
\examples{
suma(a=4, b=6)
suma(b=7, a=3)
}
