% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_scores_data.R
\docType{data}
\name{survey_scores_data}
\alias{survey_scores_data}
\title{Demo survey data from Brazilian Marine Extractive Reserves}
\format{
A data frame with 32 rows and 4 colums
\describe{
\item{id}{string, response ID}
\item{category}{(economical, ecological, social, instatution)}
\item{score}{(integer)}
\item{max}{ (integer)}
}
}
\usage{
survey_scores_data
}
\description{
Data wth survey response ID, category indicator, response score, and the maximum score
}
\keyword{datasets}
