% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{This function returns the train set of the dataframe}
\usage{
train(df)
}
\arguments{
\item{df}{The dataset which you need its train set}
}
\value{
A dataframe
}
\description{
This function returns the train set of the dataframe
}
\details{
This function allows the users to get only the train set of the insurance dataframe
to use in the other functions' arguments. Since there is another function called "train"
in the "caret" library so always use myPack::train(df)
}
