% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_array.R
\name{validate_array}
\alias{validate_array}
\title{Validate vectors of statistical parameters (mean, sd, cv)}
\usage{
validate_array(name_arr, arr, matRef, name_mat)
}
\arguments{
\item{name_arr}{string with name of the object that represent input array to validate}

\item{arr}{object that represent input array to validate}

\item{matRef}{matrix object that represent correlation matrix/covariance matrix}

\item{name_mat}{string with name of the object that represent correlation matrix/covariance matrix}
}
\description{
This function is used to validate vectors of statistical parameters (mean, sd, cv) which are typically
the input of the functions in this package. They must be array objects without NA, NaN or infinite values and
their length must be equal to the size of the associated covariance matrix/correlation matrix (square matrix).
If these requisites are not respected, excecution will be stopped.
}
\seealso{
\code{\link{validate_corrMatrix}}

\code{\link{validate_covMatrix}}
}
\author{
Alessandro De Carlo \email{alessandro.decarlo01@universitadipavia.it}
}
