% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_corrMatrix.R
\name{validate_corrMatrix}
\alias{validate_corrMatrix}
\title{Validate input correlation matrix}
\usage{
validate_corrMatrix(corrMatStart)
}
\arguments{
\item{corrMatStart}{object that represents input correlation matrix to be validated.}
}
\description{
function used for check that input matrix is a matrix object which represents a valid correlation
matrix. If these requisites are not respected, excecution will be stopped.
}
\note{
Eigenvalues are non considered because the functions in which \emph{validate_corrMatrix} is called
can handle not PD matrices too by approximating them to nearest PD.
}
\seealso{
\code{\link[Matrix]{nearPD}}

\code{\link{validate_array}}

\code{\link{validate_covMatrix}}
}
\author{
Alessandro De Carlo \email{alessandro.decarlo01@universitadipavia.it}
}
