% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_covMatrix.R
\name{validate_covMatrix}
\alias{validate_covMatrix}
\title{Validate input covariance matrix}
\usage{
validate_covMatrix(covMatrix)
}
\arguments{
\item{covMatrix}{object that represents input covariance matrix to validate.}
}
\description{
function used for check that input matrix is a matrix object which represents a valid covariance
matrix. If these requisites are not respected, excecution will be stopped.
}
\note{
Eigenvalues are non considered because the package can handle not PD matrices too
by approximating them to nearest PD.
}
\seealso{
\code{\link{validate_array}}

\code{\link{validate_corrMatrix}}
}
\author{
Alessandro De Carlo \email{alessandro.decarlo01@universitadipavia.it}
 @seealso \code{\link[Matrix]{nearPD}}
}
