% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_groups.R
\name{water_groups}
\alias{water_groups}
\title{Water groups}
\usage{
water_groups(temperature)
}
\arguments{
\item{temperature}{water temperature in degree C, numeric variable}
}
\value{
water group based on the temperature
}
\description{
This function groups water by its temperature.
}
\examples{
water_groups(temperature = c(-1, 0, 3, 25, 42, 77, 100, 103))
water_groups(temperature = water_temp$water_temp)
}
\author{
Shuhan Song
}
