% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_temp.R
\docType{data}
\name{water_temp}
\alias{water_temp}
\title{Water Temperature Vector}
\format{
A data frame with 8 rows and 1 colums
\describe{
\item{water_temp}{numeric variable of water temperature in Celcius.}
}
}
\usage{
water_temp
}
\description{
Vector length of 8 with temperature in Celcius.
}
\keyword{datasets}
