% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groundwater_category.R
\name{gw_category}
\alias{gw_category}
\title{Groundwater Category}
\usage{
gw_category(depth)
}
\arguments{
\item{depth}{to water in meters, numeric variable}
}
\value{
category based on the depth to water
}
\description{
This function categorizes a well as "sustainable", "overdrafted", or "critically overdrafted" based on depth to water in a groundwater well
}
\author{
Annika Leiby
}
