% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyPack-package.R
\docType{package}
\name{MyPack-package}
\alias{MyPack}
\alias{MyPack-package}
\title{MyPack: My Personal Package}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
