\name{mypack-package}
\alias{mypack-package}
\alias{mypack}
\docType{package}
\title{
\packageTitle{mypack}
}
\description{
\packageDescription{mypack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mypack}
\packageIndices{mypack}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{mypack}

Maintainer: \packageMaintainer{mypack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
