% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_plot.R
\name{pred_plot}
\alias{pred_plot}
\title{Plot the predicted values vs. the actual values}
\usage{
pred_plot(pred, test)
}
\arguments{
\item{pred}{The predicted values using random forest}

\item{test}{The test set after spiliting the insurance dataset}
}
\value{
A plot
}
\description{
Plot the predicted values vs. the actual values
}
\details{
This function allows the users to plot the actual values vs. the predicted
values. Note that as the first argument you can use
predicted(train(insurance),test(insurance)), and as the second argument you can use
test(insurance)
}
