% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{predicted}
\alias{predicted}
\title{Predict the charges using Random Forest model}
\usage{
predicted(train, test)
}
\arguments{
\item{train}{The train set after spiliting the insurance data}

\item{test}{The test set after spiliting the insurance data}
}
\value{
A random forets model summary
}
\description{
Predict the charges using Random Forest model
}
\details{
This function allows the users to find the predicted values for charges using
the random forest model Note that instead if the arguments of this function
you can use train(insurance) and test(insurance).
}
